Move Jutta proto sources into component
