+++ esphome/components/jutta_proto/internal/serial/serial_connection.cpp
@@ -1,4 +1,4 @@
-#include "serial/SerialConnection.hpp"
+#include "serial_connection.hpp"
 
 #include "esphome/core/log.h"
 #include <array>
