+++ esphome/components/jutta_proto/internal/serial/serial_connection.hpp
